Abstract core if-env functionality to a separate file and reference it.
